Fix notify tests + Move controller helper funcs to util/
