+++ api/src/controller.ts
@@ -1,37 +1,15 @@
 import { Request, Response } from 'express';
-import { Queue } from 'bull';
 import { SitemapStream } from 'sitemap';
 import { createGzip } from 'zlib';
 
-import { ProjectDocumentStore, NotFoundException } from './projectdocumentstore';
-import { Validator } from './validate';
-import { IOMEGAPairedOmicsDataPlatform as ProjectDocument } from './schema';
+import { NotFoundException } from './projectdocumentstore';
 import { computeStats } from './util/stats';
 import { summarizeProject } from './summarize';
 import { ZENODO_DEPOSITION_ID } from './util/secrets';
-import { FilterFields, DEFAULT_PAGE_SIZE, Order, SearchOptions, SortFields } from './store/search';
 import { notifyNewProject } from './util/notify';
+import { getValidator, getStore, getEnrichQueue, getPendingProjectUrl, getSchema } from './util/context';
+import { validateSearchOptions } from './util/search';
 
-
-function getStore(req: Request) {
-    return req.app.get('store') as ProjectDocumentStore;
-}
-
-function getSchema(req: Request) {
-    return req.app.get('schema');
-}
-
-function getValidator(req: Request) {
-    return req.app.get('validator') as Validator;
-}
-
-function getEnrichQueue(req: Request) {
-    return req.app.get('enrichqueue') as Queue<[string, ProjectDocument]>;
-}
-
-function getPendingProjectUrl(req: Request, project_id: string) {
-    return 'https://' + req.hostname + '/pending/' + project_id;
-}
 
 export async function createProject(req: Request, res: Response) {
     const project = req.body;
@@ -89,61 +67,6 @@ export async function denyProject(req: Request, res: Response) {
     await store.denyProject(project_id);
     res.status(200);
     res.json({'message': 'Denied pending project'});
-}
-
-function checkRange(svalue: string, label: string, min: number, max: number) {
-    const value = parseInt(svalue);
-    if (Number.isNaN(value)) {
-        throw `${label} is not an integer`;
-    }
-    if (value < min || value > max ) {
-        throw `${label} must be between \`${min}\` and \`${max}\``;
-    }
-    return value;
-}
-
-function validateSearchOptions(query: any) {
-    const options: SearchOptions = {};
-
-    if (query.q) {
-        options.query = query.q;
-    }
-    if (query.fk || query.fv) {
-        if (query.fk && query.fv) {
-            if (!(query.fk in FilterFields)) {
-                throw 'Invalid `fk`';
-            }
-            options.filter = {
-                key: query.fk,
-                value: query.fv
-            };
-        } else {
-            throw 'Require both `fk` and `fv` to filter';
-        }
-    }
-    if (query.size) {
-        options.size = checkRange(query.size, 'Size', 1, 1000);
-    } else {
-        options.size = DEFAULT_PAGE_SIZE;
-    }
-    if (query.page) {
-        options.from = checkRange(query.page, 'Page', 0, 1000) * options.size;
-    } else {
-        options.from = 0;
-    }
-    if (query.sort) {
-        if (!(query.sort in SortFields)) {
-            throw 'Invalid `sort`';
-        }
-        options.sort = query.sort;
-    }
-    if (query.order) {
-        if (!(query.order in Order)) {
-            throw 'Invalid `order`, must be either `desc` or `asc`';
-        }
-        options.order = Order[query.order as 'desc' | 'asc'];
-    }
-    return options;
 }
 
 export async function listProjects(req: Request, res: Response) {
