+++ api/src/util/context.ts
@@ -1,0 +1,25 @@
+import { Request } from 'express';
+import { Queue } from 'bull';
+import { ProjectDocumentStore } from '../projectdocumentstore';
+import { Validator } from '../validate';
+import { IOMEGAPairedOmicsDataPlatform as ProjectDocument } from '../schema';
+
+export function getStore(req: Request) {
+    return req.app.get('store') as ProjectDocumentStore;
+}
+
+export function getSchema(req: Request) {
+    return req.app.get('schema');
+}
+
+export function getValidator(req: Request) {
+    return req.app.get('validator') as Validator;
+}
+
+export function getEnrichQueue(req: Request) {
+    return req.app.get('enrichqueue') as Queue<[string, ProjectDocument]>;
+}
+
+export function getPendingProjectUrl(req: Request, project_id: string) {
+    return 'https://' + req.hostname + '/pending/' + project_id;
+}
